[PATCH] Linux-0.99.14 (November 28, 1993)
Original Changelog:

CHANGES since 0.99 patchlevel 13:

 - new kernel source layout: drivers separated
 - lots of networking bugs fixed, and new network card drivers (Alan Cox,
   Donald Becker &co)
 - sound driver added to the default source distribution (Hannu
   Savolainen)
 - updated SCSI driver code (Eric Youngdale, Drew Eckhardt &co)
 - readonly OS/2 filesystem support (HPFS) added (Chris Smith)
 - NTP support (Philip Gladstone, Torsten Duwe, ??)
 - fixed 16MB swap-area limit
 - lots of minor cleanups, buxfixes etc.

+++ fs/ext2/acl.c
@@ -9,9 +9,11 @@
  * second extended file system.
  */
 
-#include <linux/sched.h>
-#include <linux/ext2_fs.h>
 #include <linux/errno.h>
+#include <linux/fs.h>
+#include <linux/ext2_fs.h>
+#include <linux/kernel.h>
+#include <linux/sched.h>
 #include <linux/stat.h>
 
 /*
@@ -21,7 +23,7 @@
  */
 int ext2_permission (struct inode * inode, int mask)
 {
-	int mode = inode->i_mode;
+	unsigned short mode = inode->i_mode;
 
 	/* Special case, access is always granted for root */
 	if (suser ())
